Fix sphinx docs (#92)
* Fix pygments lexers

`log` (used in beam_search_generator.rst) is not a valid lexer. `bash` is for
shell scripts. We are presenting interactive sessions and should use `console`.

* Add .ipynb_checkpoints to sphinx exclude_patterns

This is required: https://media.readthedocs.org/pdf/nbsphinx/latest/nbsphinx.pdf
Otherwise we will get a lot of warnings if the environment contains
ipynb_checkpoints..

* Declare development requirements

This may not be exhaustive yet, but are the requirements needed on my machine to
build the documentation.

* Make CI deploy stage fail if there are any Sphinx warnings

* Add 'release' dependency to 'docs' in Makefile

'docs' depends on .zip archives created in the 'release' step

* Declare word embedding evaluation extended_results.ipynb as orphan

Sphinx won't warn about orphan elements not being included in the TOC.

* Fix docstrings causing sphinx warnings

* Install ipython from conda-forge

Workaround https://github.com/ContinuumIO/anaconda-issues/issues/1430