Transform refactor (#282)
* Organize corpora by name instead of initial usage

Text corpora can be used for different tasks. The current API mandates the use
for a particular task by coupling the raw dataset representation with the the
transforms. This re-organizes the code declaring the different datasets as a
preparation for refactoring the API.

* Remove LM Dataset/Stream in favor of equivalent batchify

Before this change, the ability to transform a dataset to BPTT batches required
a dataset or stream to subclass a LanguageModelDataset / LanguageModelStream,
imposing restrictions on the use of the Dataset (such as enforcing flattening).

* Introduce EmbeddingCenterContextBatchify API alike changed BPTT API

* Update embedding training script

* Update WikiText defaults to match LanguageModelDataset defaults

* Update language model scripts

* Add stream_identity_wrappers fixture to tests

All DataStreams included in Gluon-NLP should support being wrapped by one of the
wrappers returned by this test fixture. When writing a test to test some Stream,
make sure to parameterize it by stream_identity_wrappers so that the stream is
tested with all possible stream wrappers.

* Make CorpusBatchify return a gluon.data.Dataset instead of ndarray

* Update language_model.ipynb

* Update sentiment_analysis.ipynb

* Update embedding.evaluation

* Update word_embedding_training.ipynb

* Add test_gbw.py

* Improve doc

* Add missing class to doc